+++ lib/src/features/home/homepage.dart
@@ -1,12 +1,16 @@
 import 'dart:async';
 
 import 'package:anilist/anilist.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
+import 'package:flutter_animate/flutter_animate.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:gap/gap.dart';
+import 'package:kurumi/src/core/core.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 import 'package:upgrader/upgrader.dart';
 
-import 'package:kurumi/src/core/utils/utils.functions.dart';
 import 'package:kurumi/src/features/activity/activity_screen.dart';
 import 'package:kurumi/src/features/anime/anime_screen.dart';
 import 'package:kurumi/src/features/discover/discover.dart';
@@ -19,6 +23,97 @@ final currentIndex = StateProvider<int>((ref) => 0);
 final navBarVisibilityProvider = StateProvider<bool>((ref) => true);
 final animeTabProvider = StateProvider<int>((ref) => 0);
 final mangaTabProvider = StateProvider<int>((ref) => 0);
+
+class WhatsNewWidget extends StatefulWidget {
+  const WhatsNewWidget({super.key});
+
+  @override
+  State<WhatsNewWidget> createState() => _WhatsNewWidgetState();
+}
+
+class _WhatsNewWidgetState extends State<WhatsNewWidget> {
+  final List changes = [
+    {
+      "id": 1,
+      "title": "3x3 Maker for Anime & Manga",
+      "detail": [
+        "Added a feature to create anime/manga 3x3 grids.",
+        "Accessible by clicking the 9-dot icon on the profile screen."
+      ],
+    },
+  ];
+
+  final colorList = [
+    Colors.green,
+    Colors.orange,
+    Colors.lightBlue,
+    Colors.pinkAccent
+  ];
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 300,
+      margin: const EdgeInsets.all(7),
+      decoration: BoxDecoration(
+        color: Colors.black.withOpacity(0.8),
+        border: Border.all(color: kWhiteColor.withOpacity(0.4)),
+        borderRadius: BorderRadius.circular(16),
+      ),
+      child: Column(
+        children: [
+          ListTile(
+            title: const Text("What's New"),
+            subtitle: const Text('v1.13.0'),
+            trailing: GestureDetector(
+              onTap: () => Navigator.pop(context),
+              child: const Icon(CupertinoIcons.clear),
+            ),
+          ),
+          for (final part in changes)
+            Padding(
+              padding: const EdgeInsets.only(left: 15, right: 20),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  Text(
+                    part['title'],
+                    style: Poppins(
+                      fontWeight: FontWeight.w600,
+                      fontSize: 18,
+                      color: colorList[part['id'] % 4],
+                    ),
+                  ),
+                  const Gap(7),
+                  ...(part["detail"] as List).map(
+                    (e) => Padding(
+                      padding: const EdgeInsets.symmetric(vertical: 3),
+                      child: Row(
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        children: [
+                          const Padding(
+                            padding: EdgeInsets.only(top: 7),
+                            child: Icon(Icons.square, size: 8),
+                          ),
+                          const Gap(10),
+                          Expanded(
+                            child: Text(
+                              e.toString(),
+                              style: Poppins(),
+                            ),
+                          ),
+                        ],
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            )
+        ],
+      ),
+    );
+  }
+}
 
 class HomePage extends ConsumerStatefulWidget {
   const HomePage({super.key});
@@ -38,6 +133,20 @@ class _HomePageState extends ConsumerState<HomePage> {
     isNavVisible = true;
     super.initState();
     initialize();
+    whatsNewWidget(context);
+  }
+
+  Future<void> whatsNewWidget(BuildContext context) async {
+    final pref = await SharedPreferences.getInstance();
+    await Future.delayed(2.seconds);
+    if (pref.getBool('whatsnew@0.13.0') != true) {
+      showModalBottomSheet(
+        context: context,
+        backgroundColor: kTransparentColor,
+        builder: (context) => const WhatsNewWidget(),
+      );
+      await pref.setBool('whatsnew@0.13.0', true);
+    }
   }
 
   Future<void> initialize() async {